requirements.txt
torch==2.7.0+cu118
transformers==4.51.3
tensorboard==2.19.0
pandas==2.2.3
datasets==3.0.0
modelscope==1.26.0
peft==0.15.2

python-docx==1.1.2
PyPDF2==3.0.1
faiss-cpu==1.11.0
sentence_transformers==4.1.0

vllm>=0.8.5